Silence uint32 to uint8 cast warning

+++ libs/kimath/src/md5_hash.cpp
@@ -267,14 +267,14 @@ void MD5_HASH::md5_final( MD5_CTX *ctx, uint8_t hash[] )
 
     // Append to the padding the total message's length in bits and transform.
     DBL_INT_ADD(ctx->bitlen[0],ctx->bitlen[1],8 * ctx->datalen);
-    ctx->data[56] = ctx->bitlen[0];
-    ctx->data[57] = ctx->bitlen[0] >> 8;
-    ctx->data[58] = ctx->bitlen[0] >> 16;
-    ctx->data[59] = ctx->bitlen[0] >> 24;
-    ctx->data[60] = ctx->bitlen[1];
-    ctx->data[61] = ctx->bitlen[1] >> 8;
-    ctx->data[62] = ctx->bitlen[1] >> 16;
-    ctx->data[63] = ctx->bitlen[1] >> 24;
+    ctx->data[56] = static_cast<uint8_t>( ctx->bitlen[0] );
+    ctx->data[57] = static_cast<uint8_t>(ctx->bitlen[0] >> 8 );
+    ctx->data[58] = static_cast<uint8_t>(ctx->bitlen[0] >> 16 );
+    ctx->data[59] = static_cast<uint8_t>(ctx->bitlen[0] >> 24 );
+    ctx->data[60] = static_cast<uint8_t>(ctx->bitlen[1] );
+    ctx->data[61] = static_cast<uint8_t>(ctx->bitlen[1] >> 8 );
+    ctx->data[62] = static_cast<uint8_t>(ctx->bitlen[1] >> 16 );
+    ctx->data[63] = static_cast<uint8_t>(ctx->bitlen[1] >> 24 );
     md5_transform(ctx,ctx->data);
 
     // Since this implementation uses little endian byte ordering and MD uses big endian,
